down to n+1 calls
side effect: c++17 is now required. the preparation of the argument string is
very hard in eralier versions as the string needs to be deducted from a type.
template metaprogramming supports mostly transformation type->type or
value->value, though not between... C++17 is easier there as with it nearly
everything can be computed at compile time.